# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_service.CreateInstanceRequest:
    properties:
      actions:
        items:
          type: integer
        type: array
      data_config:
        items:
          type: integer
        type: array
      data_endpoint:
        type: string
      description:
        type: string
      handler_key:
        type: string
      name:
        type: string
      required_permission:
        type: string
      scope:
        type: string
      screen_key:
        type: string
      slot_data:
        items:
          type: integer
        type: array
      template_id:
        type: string
    required:
    - name
    - screen_key
    - template_id
    type: object
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_service.CreateTemplateRequest:
    properties:
      definition:
        items:
          type: integer
        type: array
      description:
        type: string
      name:
        type: string
      pattern:
        type: string
    required:
    - definition
    - name
    - pattern
    type: object
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_service.LinkScreenRequest:
    properties:
      is_default:
        type: boolean
      resource_id:
        type: string
      resource_key:
        type: string
      screen_key:
        type: string
      screen_type:
        type: string
    required:
    - resource_id
    - resource_key
    - screen_key
    - screen_type
    type: object
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_service.UpdateInstanceRequest:
    properties:
      actions:
        items:
          type: integer
        type: array
      data_config:
        items:
          type: integer
        type: array
      data_endpoint:
        type: string
      description:
        type: string
      handler_key:
        type: string
      name:
        type: string
      required_permission:
        type: string
      scope:
        type: string
      screen_key:
        type: string
      slot_data:
        items:
          type: integer
        type: array
      template_id:
        type: string
    type: object
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_service.UpdateTemplateRequest:
    properties:
      definition:
        items:
          type: integer
        type: array
      description:
        type: string
      name:
        type: string
      pattern:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Clean administration API for EduGo
  title: EduGo API Admin New
  version: "1.0"
paths:
...
  /screen-config/instances:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List screen instances
      tags:
      - screen-config
    post:
      consumes:
      - application/json
      parameters:
      - description: Instance data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_service.CreateInstanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a screen instance
      tags:
      - screen-config
  /screen-config/instances/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Instance ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a screen instance
      tags:
      - screen-config
    get:
      consumes:
      - application/json
      parameters:
      - description: Instance ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a screen instance by ID
      tags:
      - screen-config
    put:
      consumes:
      - application/json
      parameters:
      - description: Instance ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Instance update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_service.UpdateInstanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a screen instance
      tags:
      - screen-config
  /screen-config/instances/key/{key}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Instance key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a screen instance by key
      tags:
      - screen-config
  /screen-config/resolve/key/{key}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Screen key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resolve a screen configuration by key
      tags:
      - screen-config
  /screen-config/resource-screens:
    post:
      consumes:
      - application/json
      parameters:
      - description: Link data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_service.LinkScreenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Link a screen instance to a resource
      tags:
      - screen-config
  /screen-config/resource-screens/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Resource-Screen link ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unlink a screen from a resource
      tags:
      - screen-config
  /screen-config/resource-screens/{resourceId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: resourceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get screens linked to a resource
      tags:
      - screen-config
  /screen-config/templates:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List screen templates
      tags:
      - screen-config
    post:
      consumes:
      - application/json
      parameters:
      - description: Template data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_service.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a screen template
      tags:
      - screen-config
  /screen-config/templates/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Template ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a screen template
      tags:
      - screen-config
    get:
      consumes:
      - application/json
      parameters:
      - description: Template ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a screen template by ID
      tags:
      - screen-config
    put:
      consumes:
      - application/json
      parameters:
      - description: Template ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Template update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_service.UpdateTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a screen template
      tags:
      - screen-config
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT token. Example: "Bearer
      eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"